+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.4.0
+
+- Change default render behavior to only show a limited number charts and add a Tab to all charts.
+  This avoids the page being overloaded by charts. The number of charts on the object page is configurable
+- Add dedicated page to show all charts for a given host/service that can be used in a Dashboard
+- Add check_intervall to PerfdataRequest
+- Fix some missing translation hooks
+
 ## v0.3.2
 
 - Fix custom timeranges not being applied